Created API for ArachneRobot class
The addition of the API enables multiple classes to modify the core
robot behaviour in each mode without overriding ArachneRobot. This means
that secondary modules such as the complex motion module can be
developed separately, despite needing to have actions performed on the
initialisation and execution of different modes - in the case of the
complex motion module, for instance, PID controllers must be disabled
when the robot is disabled to prevent unexpected behaviour upon
re-enabling.