Removed duplicate sensor settings
Originally, sensors could be created for rate or displacement by either
creating a generic sensor and setting a variable, or by calling a
constructor from a dedicated class for rate or displacement.
The first method has been removed, and the generic class made into an
abstract superclass for the two dedicated classes for the relevant
sensors.